remove circular dependencies from abstract-error
